Fixed inconsistent elapsed time frame-rate
Former-commit-id: 5a764e3d3b519bad3f70a2e59ae9ba0bfa1bb1c1

+++ include/window.h
@@ -17,7 +17,7 @@ void window_refresh(struct window *window);
 void window_unuse(struct window *window);
 
 // Utilities
-double window_elapsed_seconds(struct window *window);
+float window_elapsed_seconds(struct window *window);
 void window_framebuffer_size(struct window *window, int *width, int *height);
 
 // Closing
